requirements.txt
Flask==3.0.0
flask-cors==4.0.0
PyJWT==2.8.0
gunicorn==21.2.0
python-dotenv==1.0.0
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1
